+++ src/app/static/src/tests/downloadResults/actions.test.ts
@@ -7,7 +7,6 @@ import {
 import {actions} from "../../app/store/downloadResults/actions";
 import {DOWNLOAD_TYPE} from "../../app/types";
 import {DownloadStatusResponse} from "../../app/generated";
-import {switches} from "../../app/featureSwitches";
 
 const RunningStatusResponse: DownloadStatusResponse = {
     id: "db0c4957aea4b32c507ac02d63930110",
@@ -726,7 +725,6 @@ describe(`download Results actions`, () => {
     });
 
     it("can submit comparison download request, commits and starts polling", async () => {
-        switches.comparisonOutput = true;
         const commit = jest.fn();
         const dispatch = jest.fn();
         const downloadId = {downloadId: "1"};
@@ -763,6 +761,31 @@ describe(`download Results actions`, () => {
         await actions.prepareComparisonOutput({commit, state} as any);
         expect(mockAxios.history.post.length).toBe(0);
         expect(commit.mock.calls.length).toBe(0);
+    });
+
+    it("download comparison report", async () => {
+        const commit = jest.fn();
+
+        const root = mockRootState({
+            modelCalibrate: mockModelCalibrateState({calibrateId: "calibrate1"}),
+        });
+
+        const state = mockDownloadResultsState({
+            comparison: mockDownloadResultsDependency({downloadId: "1"})
+        });
+
+        await actions.downloadComparisonReport({commit, state, rootState: root} as any,);
+        expect(commit.mock.calls.length).toBe(1);
+        expect(commit.mock.calls[0][0]["type"]).toBe("errors/ErrorAdded")
+        expect(commit.mock.calls[0][0]["payload"]).toEqual(
+            {
+                detail: "Could not parse API response. Please contact support.",
+                error: "MALFORMED_RESPONSE"
+            }
+        )
+        expect(mockAxios.history.get.length).toBe(1);
+
+        expect(mockAxios.history.get[0]["url"]).toBe("download/result/1");
     });
 
     it("prepare comparison does not do anything if fetchingDownloadId is set", async () => {
@@ -918,7 +941,6 @@ describe(`download Results actions`, () => {
     });
 
     it("does not start polling for comparison output status when submission is unsuccessful", async () => {
-        switches.comparisonOutput = true;
         const commit = jest.fn();
         const dispatch = jest.fn();